+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [5.12.0] - 2025-07-08
-- added `GET /resources` endpoint to retrieve resources with optional filtering and pagination
-- added `ResourceCollector` class for retrieving resources using Cloud Custodian filters 
+- added `GET /resources` and `GET /resources/arn/{arn}` endpoints to retrieve resources with optional filtering and pagination.
+- added `ResourceCollector` class for retrieving resources using Cloud Custodian filters.
 
 ## [5.11.0] - 2025-06-13
 - updated `modular-sdk` version to `7.1.0`